requirements.txt
streamlit>=1.35.0
torch>=2.1.0
diffusers>=0.27.0
transformers>=4.40.0
trimesh>=4.0.5
scipy>=1.12.0
pyvista>=0.43.0
pyngrok>=7.0.0
streamlit-pyvista>=0.0.5
stpyvista>=0.0.10
python-dotenv>=1.0.1
google-generativeai>=0.3.2